reafctor :
- test 클래스 분리
- Rank : getCount 메소드 삭제
- WinRank : Lottos 기능 분리
- ResultView : Lotto,Lottos print 분리
